pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fmriflow"
version = "2.0.0a3"
description = "fMRIflow — Module-based neuroscience encoding model pipeline"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
cloud = ["cottoncandy>=0.2", "boto3>=1.26"]
viz = ["pycortex>=1.2", "matplotlib>=3.7"]
ml = ["scikit-learn>=1.2"]
himalaya = ["himalaya>=0.4", "scikit-learn>=1.2"]
nlp = ["transformers>=4.28", "gensim>=4.3"]
audio = ["librosa>=0.10", "soundfile>=0.12"]
video = ["opencv-python>=4.7"]
bids = ["nibabel>=5.0"]
flatten = ["autoflatten>=0.1.0"]
frontend = ["fastapi>=0.109", "uvicorn>=0.27", "websockets>=12.0"]
all = ["fmriflow[cloud,viz,ml,nlp,himalaya,audio,video,bids,frontend,flatten]"]
dev = ["fmriflow[all]", "pytest>=7.3", "pytest-cov>=4.1", "h5py>=3.8"]

[project.scripts]
fmriflow = "fmriflow.cli:main"

# Built-in modules registered via entry points
[project.entry-points."fmriflow.stimulus_loaders"]
textgrid = "fmriflow.modules.stimulus_loaders.textgrid:TextGridStimulusLoader"
audio = "fmriflow.modules.stimulus_loaders.audio:AudioStimulusLoader"
video = "fmriflow.modules.stimulus_loaders.video:VideoStimulusLoader"

[project.entry-points."fmriflow.response_loaders"]
bids = "fmriflow.modules.response_loaders.bids:BidsResponseLoader"
cloud = "fmriflow.modules.response_loaders.cloud:CloudResponseLoader"
local = "fmriflow.modules.response_loaders.local:LocalResponseLoader"

[project.entry-points."fmriflow.feature_extractors"]
numwords = "fmriflow.modules.feature_extractors.basic:NumWordsExtractor"
numletters = "fmriflow.modules.feature_extractors.basic:NumLettersExtractor"
numphonemes = "fmriflow.modules.feature_extractors.basic:NumPhonemesExtractor"
word_length_std = "fmriflow.modules.feature_extractors.basic:WordLengthStdExtractor"
english1000 = "fmriflow.modules.feature_extractors.histograms:English1000Extractor"
letters = "fmriflow.modules.feature_extractors.histograms:LetterHistogramExtractor"
phonemes = "fmriflow.modules.feature_extractors.histograms:PhonemeHistogramExtractor"
word2vec = "fmriflow.modules.feature_extractors.embeddings:Word2VecExtractor"
bert = "fmriflow.modules.feature_extractors.embeddings:BERTExtractor"
fasttext = "fmriflow.modules.feature_extractors.embeddings:FastTextExtractor"
mel_spectrogram = "fmriflow.modules.feature_extractors.audio:MelSpectrogramExtractor"
rms_energy = "fmriflow.modules.feature_extractors.audio:RMSEnergyExtractor"
luminance = "fmriflow.modules.feature_extractors.visual:LuminanceExtractor"
motion_energy = "fmriflow.modules.feature_extractors.visual:MotionEnergyExtractor"

[project.entry-points."fmriflow.feature_sources"]
compute = "fmriflow.modules.feature_sources.compute:ComputeSource"
filesystem = "fmriflow.modules.feature_sources.filesystem:FilesystemSource"
cloud = "fmriflow.modules.feature_sources.cloud:CloudSource"

[project.entry-points."fmriflow.preparers"]
default = "fmriflow.modules.preparers.default:DefaultPreparer"
pre_prepared = "fmriflow.modules.preparers.pre_prepared:PreparedDataLoader"

[project.entry-points."fmriflow.models"]
bootstrap_ridge = "fmriflow.modules.models.ridge:BootstrapRidgeModel"
himalaya_ridge = "fmriflow.modules.models.himalaya:HimalayaRidgeModel"
banded_ridge = "fmriflow.modules.models.himalaya:BandedRidgeModel"
multiple_kernel_ridge = "fmriflow.modules.models.himalaya:MultipleKernelRidgeModel"

[project.entry-points."fmriflow.reporters"]
metrics = "fmriflow.modules.reporters.metrics:MetricsReporter"
flatmap = "fmriflow.modules.reporters.flatmap:FlatmapReporter"
weights = "fmriflow.modules.reporters.weights:WeightsReporter"
histogram = "fmriflow.modules.reporters.histogram:HistogramReporter"
webgl = "fmriflow.modules.reporters.webgl:WebGLReporter"

[tool.hatch.build.targets.wheel]
packages = ["fmriflow"]

# Force-include the built frontend bundle. Hatch's default file
# selection skips it in some sandboxes (e.g. docker build context
# without a .gitignore present), even though it lives under the
# packaged tree. force-include is unconditional.
[tool.hatch.build.targets.wheel.force-include]
"fmriflow/server/static" = "fmriflow/server/static"
